// data/github/current/ui/packages/issue-viewer/components/sections/__tests__/__generated__/OptionsSectionTestQuery.graphql.ts
export type OptionsSectionTestQuery$variables = Record<PropertyKey, never>;
export type OptionsSectionTestQuery$data = {
  readonly node: {
    readonly " $fragmentSpreads": FragmentRefs<"IssueSidebarPrimaryQuery">;
  } | null | undefined;
};
export type OptionsSectionTestQuery = {
  response: OptionsSectionTestQuery$data;
  variables: OptionsSectionTestQuery$variables;
};